CTI_Code/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
neo4j==5.23.1
python-dotenv==1.0.1
